Revert "Random optimization"
This reverts commit 6c7f910fa89f874ac8175590610e6d50751bcd46.

+++ Exiled.API/Features/Spawn/DynamicSpawnPoint.cs
@@ -28,7 +28,7 @@ namespace Exiled.API.Features.Spawn
         public SpawnLocationType Location { get; set; }
 
         /// <inheritdoc/>
-        public override int Chance { get; set; }
+        public override float Chance { get; set; }
 
         /// <inheritdoc/>
         [YamlIgnore]
